refactor(gateway): extract the gateway manager into internal/gateway
Move-only: relocate gateway_manager.go (+ test) into a new internal/gateway
context as manager.go. It owns the lifecycle of the local gateway - container
or binary - plus the PID registry, and has no business sitting in a catch-all.

Only the package clause and container.go's import change.
