+++ internal/gateway/manager.go
@@ -1,4 +1,4 @@
-package services
+package gateway
 
 import (
 	"archive/tar"
