+++ internal/gateway/manager_test.go
@@ -1,4 +1,4 @@
-package services
+package gateway
 
 import (
 	"os"
